change folder structure and endpoints to support versioning
