+++ src/api/router.js
@@ -1,14 +1,12 @@
 import { health } from './health/index.js'
-import { initiateUpload } from './initiate/index.js'
-import { uploadCallback } from './callback/index.js'
-import { metadataRoute } from './metadata/index.js'
+import { uploadCallback } from './v1/callback/index.js'
+import { metadataRoute } from './v1/metadata/index.js'
 
 const router = {
   plugin: {
     name: 'Router',
     register: async (server) => {
       await server.register([health])
-      server.route(initiateUpload)
       server.route(uploadCallback)
       server.route(metadataRoute)
     }
