+++ test/integration/narrow/api/callback.test.js
@@ -25,7 +25,7 @@ afterAll(async () => {
   config.set('mongo.collections.uploadMetadata', originalCollection)
 })
 
-describe('POST to the /callback route', async () => {
+describe('POST to the /api/v1/callback route', async () => {
   server = await createServer()
   await server.initialize()
 
@@ -33,7 +33,7 @@ describe('POST to the /callback route', async () => {
     test('should save a document into the collection', async () => {
       const response = await server.inject({
         method: 'POST',
-        url: '/callback',
+        url: '/api/v1/callback',
         payload: mockMetadataPayload
       })
 
@@ -51,7 +51,7 @@ describe('POST to the /callback route', async () => {
     test('should fail validation', async () => {
       const response = await server.inject({
         method: 'POST',
-        url: '/callback',
+        url: '/api/v1/callback',
         payload: {}
       })
 
@@ -72,7 +72,7 @@ describe('POST to the /callback route', async () => {
 
       const response = await server.inject({
         method: 'POST',
-        url: '/callback',
+        url: '/api/v1/callback',
         payload: mockMetadataPayload
       })
 
@@ -88,7 +88,7 @@ describe('POST to the /callback route', async () => {
 
       const response = await server.inject({
         method: 'POST',
-        url: '/callback',
+        url: '/api/v1/callback',
         payload: mockMetadataPayload
       })
